% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFigs.R
\name{loadFigs}
\alias{loadFigs}
\title{loadFigs}
\usage{
loadFigs()
}
\description{
Tool for Loading Calculations and Pretty Figure Production
}
\details{
Shiny application that uses user-uploaded pollutant and flow datasets to calculate observed and TMDL loadings and make it easy to produce publication-ready dot plots, bar plots, and load duration curves.
}
