% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foresiteR.R
\name{foresiteR}
\alias{foresiteR}
\title{foresiteR}
\usage{
foresiteR()
}
\description{
Water Quality Assessment Planning Tool
}
\details{
Shiny application that uses the most recent IR assessment information and historical site impairments to inform visualization of monitoring locations and sampling data in the Water Quality Portal.
}
